Seats Arduino System
This is the script that ran on the Arduino board. It monitors all of the hardware/sensors & then pushes that data through serial communication to the server running a local Python script. After the Python script receives the data, it parses the information & pushes it to our web server. We were using PHP for the primary backend language & all the data was stored as JSON objects. Refer to the PHP files for further information on how that worked.